Fix for issue 1: OS detection doesn't work for non-Windows
The code was consistent with the documentation, which states that OS X will be identified with Capabilities.os == "MacOS". However, that information is incorrect. In AIR 1.0 (on my Mac which has OS X 10.4.11) the Capabilities.os value is "Mac OS 10.4.9".

Also added a catch-all "else" clause to the condition. Right now only Windows and Mac OS X are supported on AIR, but that won't be the case forever.